# Conflicts:
#	app.json
#	config/config.ini.example
#	static/cover.png

+++ pogom/models.py
@@ -2,11 +2,10 @@
 # -*- coding: utf-8 -*-
 
 import logging
-import os
 import calendar
-from peewee import Model, SqliteDatabase, InsertQuery,\
-                   IntegerField, CharField, DoubleField, BooleanField,\
-                   DateTimeField, OperationalError, create_model_tables, fn
+from peewee import SqliteDatabase, InsertQuery, \
+    IntegerField, CharField, DoubleField, BooleanField, \
+    DateTimeField, fn
 from playhouse.flask_utils import FlaskDB
 from playhouse.pool import PooledMySQLDatabase
 from playhouse.shortcuts import RetryOperationalError
@@ -33,13 +32,16 @@ class MyRetryDB(RetryOperationalError, PooledMySQLDatabase):
 def init_database(app):
     if args.db_type == 'mysql':
         log.info('Connecting to MySQL database on %s:%i', args.db_host, args.db_port)
+        connections = args.db_max_connections
+        if hasattr(args, 'accounts'):
+            connections *= len(args.accounts)
         db = MyRetryDB(
             args.db_name,
             user=args.db_user,
             password=args.db_pass,
             host=args.db_host,
             port=args.db_port,
-            max_connections=args.db_max_connections,
+            max_connections=connections,
             stale_timeout=300)
     else:
         log.info('Connecting to local SQLite database')
@@ -174,7 +176,7 @@ class Pokemon(BaseModel):
         query = (Pokemon
                  .select()
                  .where((Pokemon.pokemon_id == pokemon_id) &
-                        (Pokemon.disappear_time > datetime.utcfromtimestamp(last_appearance/1000.0))
+                        (Pokemon.disappear_time > datetime.utcfromtimestamp(last_appearance / 1000.0))
                         )
                  .order_by(Pokemon.disappear_time.asc())
                  .dicts()
@@ -183,6 +185,7 @@ class Pokemon(BaseModel):
         for a in query:
             appearances.append(a)
         return appearances
+
 
 class Pokestop(BaseModel):
     pokestop_id = CharField(primary_key=True, max_length=50)
@@ -375,6 +378,20 @@ def parse_map(map_dict, step_location):
     pokestops_upserted = 0
     gyms_upserted = 0
 
+    scanned[0] = {
+        'scanned_id': str(step_location[0]) + ',' + str(step_location[1]),
+        'latitude': step_location[0],
+        'longitude': step_location[1],
+        'last_modified': datetime.utcnow(),
+    }
+
+    while True:
+        try:
+            flaskDb.connect_db()
+            break
+        except Exception as e:
+            log.warning('%s... Retrying', e)
+
     if pokemons and config['parse_pokemon']:
         notifier.notify_pkmns(pokemons)
         pokemons_upserted = len(pokemons)
@@ -388,33 +405,33 @@ def parse_map(map_dict, step_location):
         gyms_upserted = len(gyms)
         bulk_upsert(Gym, gyms)
 
-    log.info('Upserted %d pokemon, %d pokestops, and %d gyms',
-        pokemons_upserted,
-        pokestops_upserted,
-        gyms_upserted)
-
-    scanned[0] = {
-        'scanned_id': str(step_location[0])+','+str(step_location[1]),
-        'latitude': step_location[0],
-        'longitude': step_location[1],
-        'last_modified': datetime.utcnow(),
-    }
-
     bulk_upsert(ScannedLocation, scanned)
 
     clean_database()
+
+    flaskDb.close_db(None)
+
+    log.info('Upserted %d pokemon, %d pokestops, and %d gyms',
+             pokemons_upserted,
+             pokestops_upserted,
+             gyms_upserted)
 
     return True
 
 
 def clean_database():
-    flaskDb.connect_db()
     query = (ScannedLocation
-            .delete()
-            .where((ScannedLocation.last_modified <
-                (datetime.utcnow() - timedelta(minutes=30)))))
+             .delete()
+             .where((ScannedLocation.last_modified <
+                    (datetime.utcnow() - timedelta(minutes=30)))))
     query.execute()
-    flaskDb.close_db(None)
+
+    if args.purge_data > 0:
+        query = (Pokemon
+                 .delete()
+                 .where((Pokemon.disappear_time <
+                        (datetime.utcnow() - timedelta(hours=args.purge_data)))))
+        query.execute()
 
 
 def bulk_upsert(cls, data):
@@ -422,25 +439,22 @@ def bulk_upsert(cls, data):
     i = 0
     step = 120
 
-    flaskDb.connect_db()
-
     while i < num_rows:
-        log.debug('Inserting items %d to %d', i, min(i+step, num_rows))
+        log.debug('Inserting items %d to %d', i, min(i + step, num_rows))
         try:
-            InsertQuery(cls, rows=data.values()[i:min(i+step, num_rows)]).upsert().execute()
+            InsertQuery(cls, rows=data.values()[i:min(i + step, num_rows)]).upsert().execute()
         except Exception as e:
             log.warning('%s... Retrying', e)
             continue
 
-        i+=step
-
-    flaskDb.close_db(None)
+        i += step
 
 
 def create_tables(db):
     db.connect()
     db.create_tables([Pokemon, Pokestop, Gym, ScannedLocation], safe=True)
     db.close()
+
 
 def drop_tables(db):
     db.connect()
